Move CLI code to its own file
