+++ bin/electron-packager.js
@@ -2,6 +2,7 @@
 
 'use strict'
 
+// Not consts so that this file can load in Node < 4.0
 var packageJSON = require('../package.json')
 var semver = require('semver')
 if (!semver.satisfies(process.versions.node, packageJSON.engines.node)) {
@@ -10,48 +11,5 @@ if (!semver.satisfies(process.versions.node, packageJSON.engines.node)) {
   process.exit(1)
 }
 
-// Not consts so that this file can load in Node < 4.0
-var common = require('../src/common')
-var fs = require('fs')
-var packager = require('..')
-var path = require('path')
-var usage = fs.readFileSync(path.resolve(__dirname, '..', 'usage.txt')).toString()
-
-var args = common.parseCLIArgs(process.argv.slice(2))
-
-// temporary fix for https://github.com/nodejs/node/issues/6456
-var stdioWriters = [process.stdout, process.stderr]
-stdioWriters.forEach(function (stdioWriter) {
-  if (stdioWriter._handle && stdioWriter._handle.setBlocking) {
-    stdioWriter._handle.setBlocking(true)
-  }
-})
-
-function printUsageAndExit (isError) {
-  var print = isError ? console.error : console.log
-  print(usage)
-  process.exit(isError ? 1 : 0)
-}
-
-if (args.help) {
-  printUsageAndExit(false)
-} else if (args.version) {
-  if (typeof args.version !== 'boolean') {
-    console.error('--version does not take an argument. Perhaps you meant --app-version or --electron-version?\n')
-  }
-  console.log(common.hostInfo())
-  process.exit(0)
-} else if (!args.dir) {
-  printUsageAndExit(true)
-}
-
-packager(args)
-  .then(function done (appPaths) {
-    if (appPaths.length > 1) console.error('Wrote new apps to:\n' + appPaths.join('\n'))
-    else if (appPaths.length === 1) console.error('Wrote new app to', appPaths[0])
-    return true
-  }).catch(function error (err) {
-    if (err.message) console.error(err.message)
-    else console.error(err, err.stack)
-    process.exit(1)
-  })
+var cli = require('../src/cli')
+cli.run(process.argv.slice(2))
